Add a new workflow to snapshot rawhide
We want to reuse as many layers as possible to save bandwidth and SSD
writes. There are two ways to accomplish this:

One would be to use rpm-ostree's container-encapsulate to re-optimize
the layers. I'm not sure how feasible this is to do on a free runner.

The other, which is what this commit does, is to snapshot the base
image and to put all the package updates into one layer at the bottom
of the build.